Make NavigationBar noop when edge-to-edge is enabled
